use modimport instead of GLOBAL.package.path manipulation
Manipulating GLOBAL.package.path to be able to use require() nicely
messes up the package.path for all mods being loaded after this one, so
it has to be avoided.

Klei's modinfo() function does not return a value, isn't cached nor
guarded against repeated inclusion, and "dependencies" need to be
resolved manually by importing in the correct order. All a bit iffy,
but for a project this size probably okay.

Fixes #1.